+++ test/util.ts
@@ -90,4 +90,34 @@ describe('divide', () => {
 
   });
 
+  describe('TRUNCATE', () => {
+    const tests = [
+
+      // .5 cases
+      [7n, 2n, 3n],
+      [5n, 2n, 2n],
+      [-7n, 2n, -3n],
+      [-5n, 2n, -2n],
+
+      // Common cases
+      [8n, 3n, 2n],
+      [9n, 4n, 2n],
+      [-8n, 3n, -2n],
+      [-9n, 4n, -2n],
+    ];
+
+    for(const test of tests) {
+
+      it(`${test[0]} / ${test[1]} = ${test[2]}`, () => {
+
+        expect(
+          divide(test[0], test[1], Round.TRUNCATE),
+        ).to.equal(test[2]);
+
+      });
+
+    }
+
+  });
+
 });